+++ apps/web/app/(use-page-wrapper)/[user]/[type]/page.tsx
@@ -1,21 +1,16 @@
 import { withAppDirSsr } from "app/WithAppDirSsr";
 import type { PageProps } from "app/_types";
 import { generateMeetingMetadata } from "app/_utils";
-import { WithLayout } from "app/layoutHOC";
-import { cookies, headers } from "next/headers";
+import { headers, cookies } from "next/headers";
 
 import { getOrgFullOrigin } from "@calcom/features/ee/organizations/lib/orgDomains";
 
 import { buildLegacyCtx } from "@lib/buildLegacyCtx";
-import { getServerSideProps } from "@lib/org/[orgSlug]/[user]/[type]/getServerSideProps";
 
-import type { PageProps as TeamTypePageProps } from "~/team/type-view";
-import TeamTypePage from "~/team/type-view";
-import UserTypePage from "~/users/views/users-type-public-view";
-import type { PageProps as UserTypePageProps } from "~/users/views/users-type-public-view";
+import { getServerSideProps } from "@server/lib/[user]/[type]/getServerSideProps";
 
-export type OrgTypePageProps = UserTypePageProps | TeamTypePageProps;
-const getData = withAppDirSsr<OrgTypePageProps>(getServerSideProps);
+import type { PageProps as LegacyPageProps } from "~/users/views/users-type-public-view";
+import LegacyPage from "~/users/views/users-type-public-view";
 
 export const generateMetadata = async ({ params, searchParams }: PageProps) => {
   const legacyCtx = buildLegacyCtx(headers(), cookies(), params, searchParams);
@@ -23,7 +18,6 @@ export const generateMetadata = async ({ params, searchParams }: PageProps) => {
 
   const { booking, isSEOIndexable = true, eventData, isBrandingHidden } = props;
   const rescheduleUid = booking?.uid;
-
   const profileName = eventData?.profile?.name ?? "";
   const profileImage = eventData?.profile.image;
   const title = eventData?.title ?? "";
@@ -53,12 +47,13 @@ export const generateMetadata = async ({ params, searchParams }: PageProps) => {
     },
   };
 };
+const getData = withAppDirSsr<LegacyPageProps>(getServerSideProps);
 
-const Page = async (props: OrgTypePageProps) => {
-  if ((props as TeamTypePageProps)?.teamId) {
-    return <TeamTypePage {...(props as TeamTypePageProps)} />;
-  }
-  return <UserTypePage {...(props as UserTypePageProps)} />;
+const ServerPage = async ({ params, searchParams }: PageProps) => {
+  const legacyCtx = buildLegacyCtx(headers(), cookies(), params, searchParams);
+  const props = await getData(legacyCtx);
+
+  return <LegacyPage {...props} />;
 };
 
-export default WithLayout({ getLayout: null, getData, ServerPage: Page, isBookingPage: true });
+export default ServerPage;
