+++ apps/web/lib/org/[orgSlug]/[user]/getServerSideProps.tsx
@@ -1,10 +1,11 @@
-import { getServerSideProps as GSSUserPage } from "@pages/[user]";
 import type { GetServerSidePropsContext } from "next";
 
 import { getSlugOrRequestedSlug } from "@calcom/features/ee/organizations/lib/orgDomains";
 import prisma from "@calcom/prisma";
 
 import { getServerSideProps as GSSTeamPage } from "@lib/team/[slug]/getServerSideProps";
+
+import { getServerSideProps as GSSUserPage } from "@server/lib/[user]/getServerSideProps";
 
 export const getServerSideProps = async (ctx: GetServerSidePropsContext) => {
   const team = await prisma.team.findFirst({
