+++ apps/web/middleware.ts
@@ -183,6 +183,8 @@ export const config = {
     "/routing-forms/:path*",
     "/team/:path*",
     "/org/:path*",
+    "/:user/:type/",
+    "/:user/",
   ],
 };
