+++ apps/web/pages/[user]/[type]/embed.tsx
@@ -1,7 +1,16 @@
 import withEmbedSsr from "@lib/withEmbedSsr";
 
+import PageWrapper from "@components/PageWrapper";
+
 import { getServerSideProps as _getServerSideProps } from "@server/lib/[user]/[type]/getServerSideProps";
+
+import type { PageProps } from "~/users/views/users-type-public-view";
+import TypePage from "~/users/views/users-type-public-view";
 
 export const getServerSideProps = withEmbedSsr(_getServerSideProps);
 
-export { default } from "./index";
+const Type = (props: PageProps) => <TypePage {...props} />;
+
+Type.PageWrapper = PageWrapper;
+
+export default Type;
