--- apps/web/pages/[user]/[type]/index.tsx
@@ -1,12 +1,0 @@
-import PageWrapper from "@components/PageWrapper";
-
-import type { PageProps } from "~/users/views/users-type-public-view";
-import TypePage from "~/users/views/users-type-public-view";
-
-export { getServerSideProps } from "@server/lib/[user]/[type]/getServerSideProps";
-
-const Type = (props: PageProps) => <TypePage {...props} />;
-
-Type.PageWrapper = PageWrapper;
-
-export default Type;
